traceable cache
